backend/requirements.txt
Flask>=2.2.0
Flask-Cors>=3.0.0
python-dotenv>=0.20.0
Werkzeug>=2.2.0
pg8000>=1.29.0
PyJWT>=2.6.0
requests>=2.28.0
Flask-Cors==3.0.10
python-dotenv==0.21.0
pg8000==1.29.4
Werkzeug==2.2.3
PyJWT==2.6.0
cryptography==38.0.4
